requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pymupdf
pytesseract
pillow
ollama
pandas
